chore(blueprint): refactor app to use blueprint
- use flask blueprint package to refactor app
- refactor docker compose yml file
