+++ services/users/manage.py
@@ -1,9 +1,13 @@
 import unittest
 
 from flask.cli import FlaskGroup
-from project import app, db
 
-cli =  FlaskGroup(app)
+from project import create_app, db
+from project.api.models import User
+
+
+app = create_app()
+cli =  FlaskGroup(create_app=create_app)
 
 @cli.command()
 def recreate_db():
