+++ services/users/project/__init__.py
@@ -1,30 +1,29 @@
 import os
-import datetime
+
 from flask import Flask, jsonify
 from flask_sqlalchemy import SQLAlchemy
 
-app = Flask(__name__)
 
-app_settings = os.getenv('APP_SETTINGS')
-app.config.from_object(app_settings)
+# instantiate the db
+db = SQLAlchemy()
 
 
-db = SQLAlchemy(app)
+def create_app(script_info=None):
 
-class User(db.Model):
-    __tablename__ = "users"
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    username = db.Column(db.String(128), nullable=False)
-    email = db.Column(db.String(128), nullable=False)
-    active = db.Column(db.Boolean(), default=True, nullable=False)
+    # instantiate the app
+    app = Flask(__name__)
 
-    def __init__(self, username, email):
-        self.username = username
-        self.email = email
+    # set config
+    app_settings = os.getenv('APP_SETTINGS')
+    app.config.from_object(app_settings)
 
-@app.route('/users/ping', methods=['GET'])
-def ping_pong():
-    return jsonify({
-        'status': 'success',
-        'message': 'pong!'
-    })
+    # set up extensions
+    db.init_app(app)
+
+    # register blueprints
+    from project.api.users import users_blueprint
+    app.register_blueprint(users_blueprint)
+
+    # shell context for flask cli
+    app.shell_context_processor({'app': app, 'db': db})
+    return app
