+++ services/users/project/tests/test_config.py
@@ -3,9 +3,9 @@ import unittest
 
 from flask import current_app
 from flask_testing import TestCase
+from project import create_app
 
-from project import app
-
+app = create_app()
 
 class TestDevelopmentConfig(TestCase):
     def create_app(self):
